When the gfx debugger is saving shaders make sure the directory exists. Make sure the gfx debugger unpauses when the emulator is stopped. When a wad is installed make sure directories exist. For the cmake build if a header is not provided in the check_lib macro don't check for it, and assume pkg-config was supposed to work.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@6581 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Core/Src/HW/SI_DeviceGBA.cpp
@@ -134,7 +134,8 @@ void GBASockServer::Transfer(char* si_buffer)
 	else if (cmd == CMD_WRITE)
 		num_expecting = 1;
 	if (num_received != num_expecting)
-		ERROR_LOG(SERIALINTERFACE, "%x:%x:%x", (u8)cmd, num_received, num_expecting);
+		ERROR_LOG(SERIALINTERFACE, "%x:%x:%x", (u8)cmd,
+				(unsigned int)num_received, (unsigned int)num_expecting);
 #endif
 
 	si_buffer[0] = current_data[3];
